pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-tools"
version = "1.0.0"
description = "Tools for building and testing LaTeX CV"
authors = [{ name = "Dalton Luce", email = "your.email@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = ["pypdf==6.0.0", "pyspellchecker==0.7.1", "pytest==8.4.0"]

[project.optional-dependencies]
dev = ["black", "flake8", "mypy"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
